back_service/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==10.4.0
requests==2.32.3
setuptools==72.2.0
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.44.0
